🚚 Move `blueprintFormat.ts` -> `formats/blueprint/format.ts`
Also move `BLUEPRINT_CODEC` into it's own file.
