+++ src/interface/dialog/blueprintSettings.ts
@@ -1,10 +1,10 @@
 import KofiPopup from 'src/components/kofiPopup.svelte'
 import { updateAllCubeOutlines } from 'src/mods/cubeOutlineMod'
 import type { MinecraftVersion } from 'src/systems/global'
-import { updateRenderBoxPreview, updateRotationLock } from '../../blueprintFormat'
 import { defaultValues, type ExportMode } from '../../blueprintSettings'
 import BlueprintSettingsDialogSvelteComponent from '../../components/blueprintSettingsDialog.svelte'
 import { PACKAGE } from '../../constants'
+import { updateRenderBoxPreview, updateRotationLock } from '../../formats/blueprint/format'
 import { sanitizeStorageKey } from '../../util/minecraftUtil'
 import { Valuable } from '../../util/stores'
 import { SvelteDialog } from '../../util/svelteDialog'
