+++ src/interface/dialog/locatorConfig.ts
@@ -1,7 +1,7 @@
 import { registerAction } from 'src/util/moddingTools'
-import { activeProjectIsBlueprintFormat } from '../../blueprintFormat'
 import LocatorConfigDialog from '../../components/locatorConfigDialog.svelte'
 import { PACKAGE } from '../../constants'
+import { activeProjectIsBlueprintFormat } from '../../formats/blueprint/format'
 import { LocatorConfig } from '../../nodeConfigs'
 import { Valuable } from '../../util/stores'
 import { SvelteDialog } from '../../util/svelteDialog'
