+++ src/interface/panel/variants.ts
@@ -1,6 +1,6 @@
 import { registerAction, registerMenu } from 'src/util/moddingTools'
-import { BLUEPRINT_FORMAT_ID } from '../../blueprintFormat'
 import VariantsPanel from '../../components/variantsPanel.svelte'
+import { BLUEPRINT_FORMAT_ID } from '../../formats/blueprint/format'
 import { SveltePanel } from '../../util/sveltePanel'
 import { translate } from '../../util/translation'
 import { Variant } from '../../variants'
