+++ src/mods/addLocatorActionMod.ts
@@ -1,5 +1,5 @@
 import { registerMod } from 'src/util/moddingTools'
-import { activeProjectIsBlueprintFormat } from '../blueprintFormat'
+import { activeProjectIsBlueprintFormat } from '../formats/blueprint/format'
 
 registerMod({
 	id: `animated-java:add-locator-action`,
