+++ src/mods/animationControllerMod.ts
@@ -1,5 +1,5 @@
 import { createPropertySubscribable, registerMod } from 'src/util/moddingTools'
-import { activeProjectIsBlueprintFormat } from '../blueprintFormat'
+import { activeProjectIsBlueprintFormat } from '../formats/blueprint/format'
 
 registerMod({
 	id: `animated-java:animation-controller-mod`,
