+++ src/mods/animationPropertiesMod.ts
@@ -1,5 +1,5 @@
 import { registerMod } from 'src/util/moddingTools'
-import { activeProjectIsBlueprintFormat } from '../blueprintFormat'
+import { activeProjectIsBlueprintFormat } from '../formats/blueprint/format'
 import { roundToNth } from '../util/misc'
 import { translate } from '../util/translation'
