+++ src/mods/cameraNameMod.ts
@@ -1,5 +1,5 @@
 import { registerPluginMod } from 'src/util/moddingTools'
-import { activeProjectIsBlueprintFormat } from '../blueprintFormat'
+import { activeProjectIsBlueprintFormat } from '../formats/blueprint/format'
 import { sanitizeOutlinerElementName } from '../outliner/util'
 
 registerPluginMod({
