+++ src/mods/formatIconMod.ts
@@ -1,4 +1,4 @@
-import { BLUEPRINT_FORMAT_ID } from 'src/blueprintFormat'
+import { BLUEPRINT_FORMAT_ID } from 'src/formats/blueprint/format'
 import { registerMod } from 'src/util/moddingTools'
 import Icon from '../components/icon.svelte'
 import { injectSvelteComponent } from '../util/injectSvelteComponent'
