+++ src/mods/panelMod.ts
@@ -1,4 +1,4 @@
-import { activeProjectIsBlueprintFormat } from '../blueprintFormat'
+import { activeProjectIsBlueprintFormat } from '../formats/blueprint/format'
 import { registerMod } from '../util/moddingTools'
 
 registerMod({
