+++ src/systems/exporter.ts
@@ -1,5 +1,5 @@
-import { projectTargetVersionIsAtLeast, saveBlueprint } from '../blueprintFormat'
 import { blueprintSettingErrors } from '../blueprintSettings'
+import { projectTargetVersionIsAtLeast, saveBlueprint } from '../formats/blueprint/format'
 import { openBlueprintSettingsDialog } from '../interface/dialog/blueprintSettings'
 import { PROGRESS_DESCRIPTION, openExportProgressDialog } from '../interface/dialog/exportProgress'
 import { openUnexpectedErrorDialog } from '../interface/dialog/unexpectedError'
